requirements.txt
pathlib>=1.0.1
pillow>=11.3.0
pygame>=2.6.1
google-api-python-client>=2.100.0
google-auth>=2.23.0
